feat: add bookmark button in collection screen
